improve

+++ modules/task_3/zoreev_m_moore_algorithm/moore_algorithm.h
@@ -13,10 +13,10 @@
 void randomCompleteGraph(size_t size, int64_t *graph);
 
 void printGraph(size_t size, int64_t *graph);
-void printPredecessor(size_t size, size_t *predecessor);
+void printPredecessor(size_t size, uint64_t *predecessor);
 
-size_t *mooreAlgorithm(size_t size, int64_t *graph, size_t root);
+uint64_t *mooreAlgorithm(size_t size, int64_t *graph, size_t root);
 
-size_t *mooreAlgorithmParallel(size_t size, int64_t *graph, size_t root);
+uint64_t *mooreAlgorithmParallel(size_t size, int64_t *graph, size_t root);
 
 #endif // MODULES_TASK_3_ZOREEV_M_MOORE_ALGORITHM_MOORE_ALGORITHM_H_
